backend/flask_service/requirements.txt
Flask>=2.0
psycopg2-binary>=2.9
nltk>=3.6
gunicorn>=20.0

# Optional: TextBlob alternative
# textblob>=0.17.1
flask-cors>=3.0.10